--- .github/workflows/main.yml
@@ -1,21 +1,0 @@
-name: Semantic release
-
-on: push
-jobs:
-    publish:
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/checkout@v2
-            - name: Setup Node.js
-              uses: actions/setup-node@v1
-              with:
-                  node-version: 18
-            - name: Install dependencies
-              run: npm install
-            - name: Build app
-              run: npm run build
-            - name: Semantic release
-              env:
-                  GITHUB_TOKEN: ${{ secrets.GH_KEY_FREE_BEE }}
-                  NPM_TOKEN: ${{ secrets.NPM_KEY_FREE_BEE }}
-              run: npx semantic-release
