+++ .gitignore
@@ -1,3 +1,6 @@
+lib
+/lib
+
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
