+++ .storybook/main.ts
@@ -1,25 +1,16 @@
-import type { StorybookConfig } from "@storybook/react-webpack5";
-
-const config: StorybookConfig = {
-    stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
+module.exports = {
+    stories: ["../src/**/*.stories.tsx"],
     addons: [
         "@storybook/addon-links",
         "@storybook/addon-essentials",
-        "@storybook/preset-create-react-app",
-        "@storybook/addon-onboarding",
-        "@storybook/addon-interactions",
+        "@storybook/addon-styling",
+        "@storybook/addon-actions",
     ],
+    docs: {
+        autodocs: true,
+    },
     framework: {
         name: "@storybook/react-webpack5",
         options: {},
     },
-    docs: {
-        autodocs: "tag",
-    },
-    staticDirs: ["../public"],
-    features: {
-        legacyMdx1: true, // 👈 Enables MDX v1 support
-        storyStoreV7: false, // 👈 Opt out of on-demand story loading
-    },
 };
-export default config;
