+++ .storybook/preview.ts
@@ -1,15 +1,4 @@
-import type { Preview } from "@storybook/react";
-
-const preview: Preview = {
-  parameters: {
+export const parameters = {
     actions: { argTypesRegex: "^on[A-Z].*" },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-  },
+    layout: "centered",
 };
-
-export default preview;
