+++ README.md
@@ -1,5 +1,4 @@
 #React Storybook
 
-### Check how to publish your story boook react app [https://dev.to/amalv/publish-storybook-components-to-npm-using-semantic-release-and-github-actions-6ie]
-
-### Check how to publish your story boook react app [https://github.com/prateek3255/my-awesome-component-library]
+Check how to publish your story boook react app [https://dev.to/amalv/publish-storybook-components-to-npm-using-semantic-release-and-github-actions-6ie]
+Check how to publish your story boook react app [https://github.com/prateek3255/my-awesome-component-library]
