+++ package.json
@@ -1,87 +1,66 @@
 {
     "name": "free-bee",
-    "version": "0.2.0",
+    "version": "0.4.0",
+    "description": "",
     "dependencies": {
         "@emotion/react": "^11.11.1",
         "@emotion/styled": "^11.11.0",
         "@mui/material": "^5.14.3",
-        "@storybook/mdx1-csf": "^1.0.0",
-        "@testing-library/jest-dom": "^5.17.0",
-        "@testing-library/react": "^13.4.0",
-        "@testing-library/user-event": "^13.5.0",
-        "@types/jest": "^27.5.2",
-        "@types/react": "^18.2.18",
-        "@types/react-dom": "^18.2.7",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-scripts": "5.0.1",
-        "typescript": "^4.9.5",
-        "web-vitals": "^2.1.4"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "rollup -c",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
-        "storybook": "storybook dev -p 6006",
-        "build-storybook": "storybook build"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest",
-            "plugin:storybook/recommended"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+        "classnames": "^2.3.2"
     },
     "devDependencies": {
+        "@babel/preset-env": "^7.22.10",
+        "@babel/preset-react": "^7.22.5",
+        "@babel/preset-typescript": "^7.22.5",
         "@rollup/plugin-commonjs": "^25.0.4",
-        "@rollup/plugin-node-resolve": "^15.1.0",
+        "@rollup/plugin-node-resolve": "^15.2.0",
         "@rollup/plugin-typescript": "^11.1.2",
-        "@storybook/addon-essentials": "^7.2.1",
-        "@storybook/addon-interactions": "^7.2.1",
-        "@storybook/addon-links": "^7.2.1",
+        "@storybook/addon-essentials": "^7.3.1",
+        "@storybook/addon-interactions": "^7.3.1",
+        "@storybook/addon-links": "^7.3.1",
         "@storybook/addon-onboarding": "^1.0.8",
-        "@storybook/blocks": "^7.2.1",
-        "@storybook/preset-create-react-app": "^7.2.1",
-        "@storybook/react": "^7.2.1",
-        "@storybook/react-webpack5": "^7.2.1",
+        "@storybook/blocks": "^7.3.1",
+        "@storybook/react": "^7.3.1",
+        "@storybook/react-webpack5": "^7.3.1",
         "@storybook/testing-library": "^0.2.0",
-        "@types/node": "^20.5.0",
-        "autoprefixer": "^10.4.14",
-        "babel-plugin-named-exports-order": "^0.0.2",
-        "eslint-plugin-storybook": "^0.6.13",
+        "@types/react": "^18.2.20",
         "postcss": "^8.4.28",
-        "prop-types": "^15.8.1",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "rollup": "^3.28.0",
         "rollup-plugin-dts": "^5.3.1",
         "rollup-plugin-peer-deps-external": "^2.2.4",
         "rollup-plugin-postcss": "^4.0.2",
-        "semantic-release": "^21.0.7",
-        "storybook": "^7.2.1",
-        "tailwindcss": "^3.3.3",
-        "webpack": "^5.88.2"
+        "storybook": "^7.3.1",
+        "typescript": "^5.1.6"
     },
-    "files": [
-        "dist"
+    "peerDependencies": {
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0"
+    },
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "storybook": "storybook dev -p 6006",
+        "build-storybook": "storybook build",
+        "build": "rollup -c",
+        "prepublishOnly": "npm run build"
+    },
+    "keywords": [
+        "stroybook",
+        "react",
+        "typescript"
     ],
-    "main": "dist/index.js",
-    "module": "dist/index.esm.js",
-    "release": {
-        "branches": [
-            "release/main",
-            "free-bee"
-        ]
+    "main": "lib/index.js",
+    "module": "lib/index.esm.js",
+    "types": "lib/index.d.ts",
+    "author": "Rishabh Anand",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/ranand16/free-bee/issues"
+    },
+    "homepage": "https://github.com/ranand16/free-bee#readme",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/ranand16/free-bee.git"
     }
 }
