--- src/App.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <h1 className='text-6xl text-blue-500'>Welcome to React storybook. Testing Tailwind!!!</h1>
-    </div>
-  );
-}
-
-export default App;
