+++ src/components/TextButton/TextButton.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import TextButton from "../Components/TextButton";
 import { Meta } from "@storybook/react";
+import TextButton from "./TextButton";
 
 export default {
     title: "Components/Button",
