+++ src/components/TextButton/TextButton.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-
 import Button from "@mui/material/Button";
+
 type ButtonProps = { label: string };
 function TextButton({ label }: ButtonProps) {
     return (
