+++ src/index.ts
@@ -1,4 +1,4 @@
-import RatingStar from "./Components/RatingStar";
-import TextButton from "./Components/TextButton";
+import TextButton from "./components/TextButton";
+import RatingStar from "./components/RatingStar";
 
 export { TextButton, RatingStar };
