+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
     "compilerOptions": {
-        "outDir": "dist",
         "target": "es5",
+        "outDir": "lib",
         "lib": ["dom", "dom.iterable", "esnext"],
         "declaration": true,
-        "declarationDir": "dist",
+        "declarationDir": "lib",
         "allowJs": true,
         "skipLibCheck": true,
         "esModuleInterop": true,
@@ -16,11 +16,8 @@
         "resolveJsonModule": true,
         "isolatedModules": true,
         "noEmit": true,
-        "jsx": "react",
-        "types": ["node"]
+        "jsx": "react"
     },
-    "exclude": ["node_modules", "lib"],
-    "include": ["src"]
-    // "files": ["dist"],
-    // "main": "dist/index.js"
+    "include": ["src"],
+    "exclude": ["node_modules", "lib"]
 }
